+++ test/vnctools/test-vnctools-open.sh
@@ -1,30 +1,19 @@
 function test_vnctools_open_missing_display() {
     local return_code=0
-    $(vnctools-open --hostname= --localport= --remoteport= --username= ) || return_code=1
+    $(vnctools-open --hostname=  --username= ) || return_code=1
     [[ $return_code -eq 1 ]]
 }
 
 function test_vnctools_open_missing_hostname() {
     local return_code=0
-    $(vnctools-open --display= --localport= --remoteport= --username= ) || return_code=1
-    [[ $return_code -eq 1 ]]
-}
-
-function test_vnctools_open_missing_localport() {
-    local return_code=0
-    $(vnctools-open --display= --hostname= --remoteport= --username= ) || return_code=1
-    [[ $return_code -eq 1 ]]
-}
-
-function test_vnctools_open_missing_remoteport() {
-    local return_code=0
-    $(vnctools-open --display= --hostname= --localport= --username= ) || return_code=1
+    $(vnctools-open --display=  --username= ) || return_code=1
     [[ $return_code -eq 1 ]]
 }
 
 function test_vnctools_open_missing_username() {
     local return_code=0
-    $(vnctools-open --display= --hostname= --localport= --remoteport= ) || return_code=1
+    waxwing::monkey_patch_commands_to_record_command_name_only kill open ssh trap sleepf
+    $(vnctools-open --display= --hostname=  ) || return_code=1
     [[ $return_code -eq 1 ]]
 }
 
@@ -34,8 +23,6 @@ function test_vnctools_open_with_valid_args_and_default_options() {
     vnctools-open \
         --display=display \
         --hostname=hostname \
-        --localport=localport \
-        --remoteport=remoteport \
         --username=username
     actual=$(waxwing::read_pipe)
     expected=$(cat << EOF
@@ -43,11 +30,59 @@ trap clean_up; exit 1 INT
 trap echo "ERROR: \$(caller)" >&2 ERR
 ssh -CKf -o ConnectTimeout=2 username@hostname kill -9 \$(pgrep -f vnctools-x11vnc-display)
 sleep 4
-ssh -CKf -o ConnectTimeout=2 -L localport:localhost:remoteport username@hostname x11vnc \
--tag vnctools-x11vnc-display -display :display -rfbport remoteport -localhost -noshm -usepw \
+ssh -CKf -o ConnectTimeout=2 -L 5900:localhost:5900 username@hostname x11vnc \
+-tag vnctools-x11vnc-display -display :display -rfbport 5900 -localhost -noshm -usepw \
 -forever -noxdamage -snapfb -speeds dsl
 sleep 4
+open -W vnc://localhost:5900
+EOF
+)
+    [[ ${actual} == ${expected} ]]
+}
+
+function test_vnctools_open_with_valid_args_and_only_non_default_localport() {
+    waxwing::monkey_patch_commands_to_record_command_name_only kill sleep trap
+    waxwing::monkey_patch_commands_to_record_command_name_and_args open ssh
+    vnctools-open \
+        --display=display \
+        --hostname=hostname \
+        --username=username \
+        --localport=localport
+    actual=$(waxwing::read_pipe)
+    expected=$(cat << EOF
+trap
+trap
+ssh -CKf -o ConnectTimeout=2 username@hostname kill -9 \$(pgrep -f vnctools-x11vnc-display)
+sleep
+ssh -CKf -o ConnectTimeout=2 -L localport:localhost:5900 username@hostname x11vnc \
+-tag vnctools-x11vnc-display -display :display -rfbport 5900 -localhost -noshm -usepw \
+-forever -noxdamage -snapfb -speeds dsl
+sleep
 open -W vnc://localhost:localport
+EOF
+)
+    [[ ${actual} == ${expected} ]]
+}
+
+function test_vnctools_open_with_valid_args_and_only_non_default_remoteport() {
+    waxwing::monkey_patch_commands_to_record_command_name_only kill open sleep trap
+    waxwing::monkey_patch_commands_to_record_command_name_and_args ssh
+    vnctools-open \
+        --display=display \
+        --hostname=hostname \
+        --username=username \
+        --remoteport=remoteport
+    actual=$(waxwing::read_pipe)
+    expected=$(cat << EOF
+trap
+trap
+ssh -CKf -o ConnectTimeout=2 username@hostname kill -9 \$(pgrep -f vnctools-x11vnc-display)
+sleep
+ssh -CKf -o ConnectTimeout=2 -L 5900:localhost:remoteport username@hostname x11vnc \
+-tag vnctools-x11vnc-display -display :display -rfbport remoteport -localhost -noshm -usepw \
+-forever -noxdamage -snapfb -speeds dsl
+sleep
+open
 EOF
 )
     [[ ${actual} == ${expected} ]]